Refactor Duplicate Code in Navbar and Sidebar
